pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "reactorca"
version = "0.1.0"
description = "A Python CLI tool to manage a homelab CA"
authors = [
    {name = "Reactor Admin", email = "admin@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "Topic :: Security :: Cryptography",
]
# Dependencies are managed by Poetry in [tool.poetry.dependencies]

[project.scripts]
ca = "reactor_ca.main:cli"

[tool.poetry]
name = "reactorca"
version = "0.1.0"
description = "A Python CLI tool to manage a homelab CA"
authors = ["Reactor Admin"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
click = "^8.1.3"
cryptography = "^41.0.0"
pyyaml = "^6.0.0"
rich = "^13.5.0"
yamale = "^6.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.3.0"
isort = "^5.12.0"
mypy = "^1.3.0"
ruff = "^0.1.6"

[tool.poetry.scripts]
ca = "reactor_ca.main:cli"

[tool.ruff]
line-length = 118
target-version = "py312"
select = ["E", "F", "I", "N", "W", "B"]
